requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
dateparser==1.2.0
pytz==2023.3
protobuf~=6.32.0
dotenv~=0.9.9
python-dotenv~=1.1.1
timezonefinder~=8.0.0
geopy~=2.4.1